feat: Ajustes Endpoint Correos Sugeridos udistrital/sga_documentacion#291
