+++ routers/router.go
@@ -8,9 +8,8 @@
 package routers
 
 import (
-	"github.com/udistrital/sga_admisiones_mid/controllers"
-	"github.com/udistrital/sga_admisiones_mid/services"
 	"github.com/astaxie/beego"
+	"github.com/udistrital/sga_admisiones_mid/controllers"
 	"github.com/udistrital/utils_oas/errorhandler"
 )
 
@@ -39,12 +38,11 @@ func init() {
 				&controllers.LiquidacionController{},
 			),
 		),
-		beego.NSNamespace("/inscritos", // Nuevo namespace para inscritos admitidos
-		beego.NSInclude(
-			&controllers.InscritosAdmitidosController{}, // Incluir el nuevo controlador
+		beego.NSNamespace("/gestion-correos",
+			beego.NSInclude(
+				&controllers.GestionCorreosController{},
+			),
 		),
-		beego.NSRouter("/consulta", &controllers.InscritosAdmitidosController{}, "get:ConsultaInscritosController"), // Endpoint GET para la consulta
-	),
 	)
 	beego.AddNamespace(ns)
 }
